* support autoconf 2.53  * make a note that people should be using libtool 1.4d  * cut some of the more pointless bells and whistles out of the configure    script in an effort to make it more maintainable  * modularized ipc_c_tcp.c and client_crypto.o so that they're not tied    to the curses stuff.  * timezone/daylight and other FreeBSD fixes  * more effort in the configure script to detect the common    bastardizations (ahem, ports) of Berkeley DB without falling down

+++ citadel/serv_crypto.c
@@ -42,7 +42,7 @@ SSL_CTX *ssl_ctx;				/* SSL context */
 pthread_mutex_t **SSLCritters;			/* Things needing locking */
 
 static unsigned long id_callback(void) {
-	return pthread_self();
+	return (unsigned long)pthread_self();
 }
 
 void init_ssl(void)
